requirements.txt
fastapi==0.115.3
flask==3.0.3
flask-cors==4.0.1
httpx==0.27.2
pyjwt==2.9.0
uvicorn[standard]==0.32.0
